# aws/signer/v4: avoid 32KB alloc when hashing body