Determine the body length and use `io.CopyN` instead of `io.Copy`
because the latter would otherwise allocate a 32KB temporary buffer.

This provides a 4x reduction in bytes allocated by Sign and a little performance bump:

```
name                 old time/op    new time/op    delta
PresignRequest-4       38.1µs ±13%    29.6µs ± 2%  -22.42%  (p=0.008 n=5+5)
SignRequest-4          28.4µs ±14%    20.1µs ± 2%  -29.07%  (p=0.008 n=5+5)
StripExcessSpaces-4     868ns ± 2%     892ns ± 1%   +2.76%  (p=0.032 n=5+5)

name                 old alloc/op   new alloc/op   delta
PresignRequest-4       46.1kB ± 0%    13.3kB ± 0%     ~     (p=0.079 n=4+5)
SignRequest-4          42.6kB ± 0%     9.9kB ± 0%  -76.83%  (p=0.008 n=5+5)

name                 old allocs/op  new allocs/op  delta
PresignRequest-4          139 ± 0%       140 ± 0%   +0.72%  (p=0.008 n=5+5)
SignRequest-4             109 ± 0%       110 ± 0%   +0.92%  (p=0.008 n=5+5)
```

There might be other ways to tackle this. If `*request.offsetReader` implemented `WriteTo` then `io.Copy` wouldn't have to allocate the temporary buffer. However, it wasn't immediately obvious how to do that since the underlying `buf` there is an `io.ReaderSeeker`.