@xibz 2018-07-10
Hello @bpot, thank you for taking the time to construct this. I imagine the data that is being benchmarked here is small. I am curious if you benchmarked against a larger file, couple gigs, if that delta goes down. If so, I don't know if we want the `makeSha256Reader` return an error as that might cause some unexpected behavior in user's application. Let's first benchmark against a larger file and see if this makes sense to potentially pull in, especially since we are talking about `ns`.