@bpot 2018-07-10
That makes sense. To clarify I'm mostly interested in reducing the amount of memory allocated by this method in cases where the request body is relatively small. I wouldn't expect this PR to change behavior for cases where the body is larger than [32KB](https://github.com/golang/go/blob/eb9356b7a296cb3ee6dc355191cab23d29ed3427/src/io/io.go#L391).

As an example, we have an operation which updates a single field on a DynamoDB item. `makeSha256Reader` is responsible for about 66% of the space allocated by that operation.

I'll try to see if I can make this a less risky change.