@xibz 2018-07-10
@bpot - Ah, assumed all of the first column was cpu efficiency. Yea, I'll publish some feedback and see what we can do about potential breaking changes areas