@bpot 2018-07-24
@xibz rebased and updated in response to comments.