@xibz 2018-09-04
Thank you @bpot for taking the time to make those fixes. I just merged those changes in :)